# swagger.yaml
openapi: 3.0.0
info:
  title: Energy Management System API
  version: 1.0.0
  description: API documentation for the Energy Management System
servers:
  - url: /api
    description: API base URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CheckIn:
      type: object
      required:
        - user_id
        - energy_level
        - mood
        - activity_type
      properties:
        user_id:
          type: string
        energy_level:
          type: number
          minimum: 1
          maximum: 10
        mood:
          type: string
        notes:
          type: string
        activity_type:
          type: string
        social_interaction_level:
          type: number
          minimum: 1
          maximum: 10
        created_at:
          type: string
          format: date-time
    Notification:
      type: object
      required:
        - user_id
        - type
        - title
        - message
        - priority
      properties:
        user_id:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        data:
          type: object
        delivery_time:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
  /check-ins:
    post:
      summary: Create a new check-in
      tags: [Check-ins]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckIn'
      responses:
        201:
          description: Check-in created successfully
        400:
          description: Invalid request
        401:
          description: Unauthorized

    get:
      summary: List check-ins
      tags: [Check-ins]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
          description: Items per page
      responses:
        200:
          description: List of check-ins
        401:
          description: Unauthorized
...
  /notifications:
    post:
      summary: Create a new notification
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        201:
          description: Notification created successfully
        400:
          description: Invalid request
...